Transits

Transits.IntegratedLimbDark

Transits.PolynomialLimbDark

Transits.compute_Mn_coeffs!

Transits.compute_Nn_coeffs!

Transits.compute_gn

Transits.IntegratedLimbDark — Method

IntegratedLimbDark(limbdark; N=21, basis=:legendre)
IntegratedLimbDark(u; kwargs...)

Computes the time-averaged flux in the middle of an exposure by wrapping a limb darkening law limbdark with a quadrature scheme. For each time step t, N extra points are super-sampled from t-texp/2 to t+texp/2and the time-averaged flux is calculated via quadrature.

If a set of limb darkening coefficients, u, is provided, a PolynomialLimbDark law will be used by default.

Mathematical form

\[\bar{F}(t) = \frac{1}{\Delta t}\int_{t-\Delta t / 2}^{t+\Delta t / 2}{F(t')dt'}\]

where $F$ is the wrapped limb darkening law and $\Delta t$ is the exposure time.

Quadrature

The integration is approximated via Guassian quadrature

\[\frac{1}{\Delta t} \int{F(t')dt'} \approx \frac12\sum_i^N{w_i * F(\frac{\Delta t}{2}\xi_i + t)}\]

where the weights w_i and nodes ξ_i are defined by the given quadrature rule. The nodes are defined by evaluating orthogonal polynomials N times between -1 and 1. Notice the change of interval required to go from the natural bounds of the orthogonal polynomial basis, -1, 1, to the range defined by the exposure time.

The following bases are available from FastGaussQuadrature.jl. In addition, a function can be passed which calculates nodes, weights = f(N).

:legendre - Legendre polynomial base on the open (-1, 1)

:radau - Legendre polynomial base on the semi-open [-1, 1) interval

:lobatto - Legendre polynomial base on the closed [-1, 1] interval

Transits.PolynomialLimbDark — Method

PolynomialLimbDark(u::AbstractVector)

Polynomial limb darkening using analytical integrals. The length of the u vector is equivalent to the order of polynomial used; e.g., [0.2, 0.3] corresponds to quadratic limb darkening.

Mathematical form

\[I(\mu) \propto 1 - u_1(1-\mu) - u_2(1-\mu)^2 - \dots - u_N(1-\mu)^N\]

which is equivalent to the series

\[I(\mu) \propto -\sum_{i=0}^N{u_i(1-\mu)^i}\]

with the definition $u_0 \equiv -1$.

Transits.compute_Mn_coeffs! — Method

Compute the series expandsion coefficients for the M_n integrals. Need to compute from n_max - 3:n_max. Also split for the cases k^2 < 1 and k^2 ≥ 1.

Transits.compute_Nn_coeffs! — Method

Compute the series expandsion coefficients for the N_n integrals. Need to compute from n_max - 3:n_max. Only need the case k^2 < 1.

Transits.compute_gn — Method

Transform the u_n coefficients to g_n, which are coefficients in Green's basis from REF.